test(play): move the session fake to the mock surface and public-API player tests out of src
session::testing was a #[cfg(test)] fake private to kithara-play. It is now
kithara_play::mock::SessionMock next to EqualizerMock, with session(),
session_with_shape() and session_at(rate); the last replaces the
ForeignRateSession fake the sample-rate-mismatch tests declared on their own.

Every player test that needs only the public API and the mock moves to
crates/kithara-play/tests/player.rs. The tests that reach into PlayerRuntime,
PlayerCore, the phase or the prepared decoder config stay next to the code.
