+++ crates/kithara-play/Cargo.toml
@@ -140,7 +140,7 @@ hotpath = { workspace = true, optional = true }
 [dev-dependencies]
 kithara-audio = { workspace = true, features = ["mock", "probe"] }
 kithara-bufpool = { workspace = true, features = ["test-utils"] }
-kithara-play = { path = ".", default-features = false, features = ["probe"] }
+kithara-play = { path = ".", default-features = false, features = ["mock", "probe"] }
 kithara-test-utils = { path = "../kithara-test-utils", features = ["mock", "probe", "hang"] }
 
 bytes = { workspace = true }
