--- crates/kithara-play/src/mock.rs
@@ -1,1 +1,0 @@
-pub use crate::api::equalizer::EqualizerMock;
