+++ crates/kithara-play/src/mock/mod.rs
@@ -1,0 +1,5 @@
+mod session;
+
+pub use session::{SAMPLE_RATE, SessionMock, session, session_at, session_with_shape};
+
+pub use crate::api::equalizer::EqualizerMock;
