+++ crates/kithara-play/src/mock/session.rs
@@ -6,25 +6,28 @@ use std::{
 use kithara_audio::ConsumerWakeMode;
 use kithara_platform::sync::{Arc, Mutex};
 
-use super::{AllocatedSlot, Cmd, Reply, SessionBinding, SessionDispatcher, SessionSampleRate};
 use crate::{
     PlayError, SessionDuckingMode, SharedEq, SlotId, StreamShape,
     bridge::{NodeInputs, slot_channels},
+    session::{AllocatedSlot, Cmd, Reply, SessionBinding, SessionDispatcher, SessionSampleRate},
 };
 
-pub(crate) const TEST_SAMPLE_RATE: NonZeroU32 = match NonZeroU32::new(44_100) {
+/// Sample rate every `SessionMock` answers with.
+pub const SAMPLE_RATE: NonZeroU32 = match NonZeroU32::new(44_100) {
     Some(sample_rate) => sample_rate,
     None => unreachable!(),
 };
 
-struct TestSession {
+/// A session that registers players and hands out slots without a graph.
+pub struct SessionMock {
     next_player: AtomicU64,
     next_slot: AtomicU64,
     nodes: Mutex<Vec<NodeInputs>>,
     shape: Option<StreamShape>,
+    sample_rate: NonZeroU32,
 }
 
-impl<S> SessionDispatcher<S> for TestSession {
+impl<S> SessionDispatcher<S> for SessionMock {
     fn consumer_wake_mode(&self) -> ConsumerWakeMode {
         ConsumerWakeMode::RealtimeDeferred
     }
@@ -40,7 +43,9 @@ impl<S> SessionDispatcher<S> for TestSession {
                 self.nodes.lock().push(inputs);
                 Reply::SlotAllocated(AllocatedSlot::new(control, slot))
             }
-            Cmd::QuerySampleRate => Reply::SampleRate(SessionSampleRate::new(None, 44_100)),
+            Cmd::QuerySampleRate => {
+                Reply::SampleRate(SessionSampleRate::new(None, self.sample_rate.get()))
+            }
             Cmd::QueryStreamShape => Reply::StreamShape(self.shape),
             Cmd::SessionDucking => Reply::SessionDucking(SessionDuckingMode::Off),
             _ => Reply::Ok,
@@ -49,18 +54,33 @@ impl<S> SessionDispatcher<S> for TestSession {
     }
 }
 
-pub(crate) fn test_session<S>() -> SessionBinding<S> {
-    test_session_with_shape(None)
+/// A binding to a fresh `SessionMock` with no stream shape.
+#[must_use]
+pub fn session<S>() -> SessionBinding<S> {
+    session_with_shape(None)
 }
 
-pub(crate) fn test_session_with_shape<S>(shape: Option<StreamShape>) -> SessionBinding<S> {
+/// A binding to a fresh `SessionMock` answering `shape` to stream-shape queries.
+#[must_use]
+pub fn session_with_shape<S>(shape: Option<StreamShape>) -> SessionBinding<S> {
+    binding(shape, SAMPLE_RATE)
+}
+
+/// A binding to a fresh `SessionMock` running at `sample_rate` instead of `SAMPLE_RATE`.
+#[must_use]
+pub fn session_at<S>(sample_rate: NonZeroU32) -> SessionBinding<S> {
+    binding(None, sample_rate)
+}
+
+fn binding<S>(shape: Option<StreamShape>, sample_rate: NonZeroU32) -> SessionBinding<S> {
     SessionBinding::new(
-        Arc::new(TestSession {
+        Arc::new(SessionMock {
             shape,
+            sample_rate,
             next_player: AtomicU64::new(1),
             next_slot: AtomicU64::new(0),
             nodes: Mutex::default(),
         }),
-        TEST_SAMPLE_RATE,
+        sample_rate,
     )
 }
