+++ crates/kithara-play/src/player/core.rs
@@ -238,7 +238,6 @@ impl<S> PlayerRuntime<S> {
         }
     }
 }
-
 #[cfg(test)]
 mod tests {
     use std::num::NonZeroU32;
@@ -246,32 +245,19 @@ mod tests {
     use std::sync::mpsc::{RecvTimeoutError, channel};
 
     use kithara_assets::AssetStore;
-    use kithara_audio::ConsumerWakeMode;
     use kithara_decode::GaplessMode;
-    use kithara_events::{Envelope, Event};
     use kithara_platform::{CancelToken, time::Duration};
     use kithara_test_utils::kithara;
-    use kithara_warp::{StretchControls, WarpConfig};
 
     use super::*;
     use crate::{
         PlayWorkerConfig,
         bridge::PlayerCmd,
-        effects::eq::generate_log_spaced_bands,
-        player::{PlayerConfig, PlayerConfigPatch, PlayerControlSource},
+        mock,
+        player::{PlayerConfig, PlayerConfigPatch},
         resource::{ResourceConfig, ResourceSrc},
-        session::{Cmd, Reply, SessionBinding, SessionDispatcher, SessionSampleRate, testing},
         test_pools::{TestPools, pools},
     };
-
-    #[derive(Clone, Copy)]
-    enum PlayerBasicScenario {
-        AdvanceOnEmpty,
-        EngineAccessor,
-        QueueStartsEmpty,
-        SendToSlotWithoutSlot,
-        StartsPaused,
-    }
 
     fn resource_config(input: &str) -> ResourceConfig<TestPools> {
         let pools = pools();
@@ -285,32 +271,12 @@ mod tests {
         PlayWorker::new(PlayWorkerConfig::builder(pools()).build())
     }
 
-    const FOREIGN_SAMPLE_RATE: NonZeroU32 = match NonZeroU32::new(48_000) {
-        Some(sample_rate) => sample_rate,
-        None => unreachable!(),
-    };
-
-    struct ForeignRateSession;
-
-    impl SessionDispatcher<TestPools> for ForeignRateSession {
-        fn consumer_wake_mode(&self) -> ConsumerWakeMode {
-            ConsumerWakeMode::RealtimeDeferred
-        }
-
-        fn exec(&self, cmd: Cmd<TestPools>) -> Result<Reply, PlayError> {
-            match cmd {
-                Cmd::QuerySampleRate => Ok(Reply::SampleRate(SessionSampleRate::new(None, 48_000))),
-                _ => Ok(Reply::Ok),
-            }
-        }
-    }
-
     fn player() -> PlayerImpl<TestPools> {
         PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session())
+                .session(mock::session())
                 .build(),
         )
     }
@@ -442,9 +408,9 @@ mod tests {
     fn prepare_config_applies_player_gapless_mode() {
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session())
+                .session(mock::session())
                 .gapless_mode(GaplessMode::Disabled)
                 .build(),
         );
@@ -472,7 +438,7 @@ mod tests {
             .expect("the document types");
         let mut config = PlayerConfig::builder()
             .worker(worker())
-            .session(testing::test_session())
+            .session(mock::session())
             .sample_rate(NonZeroU32::new(44_100).expect("44100 is not zero"))
             .build();
         config.apply(patch);
@@ -511,9 +477,9 @@ mod tests {
         let parent_master = CancelToken::never();
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session())
+                .session(mock::session())
                 .cancel(parent_master.clone())
                 .build(),
         );
@@ -531,163 +497,6 @@ mod tests {
     }
 
     #[kithara::test]
-    #[case(PlayerBasicScenario::StartsPaused)]
-    #[case(PlayerBasicScenario::QueueStartsEmpty)]
-    #[case(PlayerBasicScenario::AdvanceOnEmpty)]
-    #[case(PlayerBasicScenario::EngineAccessor)]
-    #[case(PlayerBasicScenario::SendToSlotWithoutSlot)]
-    fn player_basic_behaviors(#[case] scenario: PlayerBasicScenario) {
-        let player = player();
-        match scenario {
-            PlayerBasicScenario::StartsPaused => {
-                assert!((player.rate() - 0.0).abs() < f32::EPSILON);
-                assert_eq!(player.status(), PlayerStatus::Unknown);
-            }
-            PlayerBasicScenario::QueueStartsEmpty => {
-                assert_eq!(player.item_count(), 0);
-            }
-            PlayerBasicScenario::AdvanceOnEmpty => {
-                player.advance_to_next_item();
-                assert_eq!(player.current_index(), 0);
-            }
-            PlayerBasicScenario::EngineAccessor => {
-                assert!(!player.engine().is_running());
-            }
-            PlayerBasicScenario::SendToSlotWithoutSlot => {
-                let result = player.send_to_slot(PlayerCmd::SetPaused(true));
-                assert!(result.is_err());
-            }
-        }
-    }
-
-    #[kithara::test]
-    fn player_pause_without_active_slot_keeps_rate_zero() {
-        let player = player();
-        player.pause();
-        assert!((player.rate() - 0.0).abs() < f32::EPSILON);
-    }
-
-    #[kithara::test]
-    fn player_volume_clamps() {
-        let player = player();
-        player.set_volume(2.0);
-        assert!((player.volume() - 1.0).abs() < f32::EPSILON);
-        player.set_volume(-1.0);
-        assert!((player.volume() - 0.0).abs() < f32::EPSILON);
-    }
-
-    #[kithara::test]
-    fn player_muted() {
-        let player = player();
-        assert!(!player.is_muted());
-        player.set_muted(true);
-        assert!(player.is_muted());
-    }
-
-    #[kithara::test]
-    fn player_crossfade_duration() {
-        let player = player();
-        assert!((player.crossfade_duration() - 1.0).abs() < f32::EPSILON);
-        player.set_crossfade_duration(3.0);
-        assert!((player.crossfade_duration() - 3.0).abs() < f32::EPSILON);
-    }
-
-    #[kithara::test]
-    fn player_prefetch_duration() {
-        let player = player();
-        assert!((player.prefetch_duration() - 3.5).abs() < f32::EPSILON);
-        player.set_prefetch_duration(8.0);
-        assert!((player.prefetch_duration() - 8.0).abs() < f32::EPSILON);
-        player.set_prefetch_duration(-1.0);
-        assert!((player.prefetch_duration() - 0.0).abs() < f32::EPSILON);
-    }
-
-    #[kithara::test]
-    fn player_events_subscribe() {
-        let player = player();
-        let mut rx = player.subscribe();
-        player.set_volume(0.5);
-        let event = rx.try_recv();
-        assert!(event.is_ok());
-    }
-
-    #[kithara::test]
-    fn player_config_custom() {
-        let config = PlayerConfig::builder()
-            .sample_rate(testing::TEST_SAMPLE_RATE)
-            .worker(worker())
-            .session(testing::test_session())
-            .crossfade_duration(2.0)
-            .prefetch_duration(5.0)
-            .default_rate(0.5)
-            .gapless_mode(GaplessMode::MediaOnly)
-            .eq_layout(generate_log_spaced_bands(5))
-            .max_slots(2)
-            .warp(
-                WarpConfig::builder()
-                    .stretch(StretchControls::new(1.0))
-                    .build(),
-            )
-            .build();
-        let player = PlayerImpl::new(config);
-        assert!((player.crossfade_duration() - 2.0).abs() < f32::EPSILON);
-    }
-
-    /// `PlayerConfig::sample_rate` is the single place the value lives before
-    /// the `EngineConfig` it configures exists. This pins that the value the
-    /// player reports back out is the exact one the engine runs with, so the
-    /// rate the owning Host hands down cannot land somewhere the engine never
-    /// reads.
-    #[kithara::test]
-    fn a_configured_sample_rate_reaches_the_engine_it_prepares() {
-        let sample_rate = NonZeroU32::new(48_000).expect("invariant: sample rate is non-zero");
-        let player = PlayerImpl::new(
-            PlayerConfig::builder()
-                .worker(worker())
-                .session(testing::test_session())
-                .sample_rate(sample_rate)
-                .build(),
-        );
-
-        assert_eq!(player.sample_rate(), sample_rate.get());
-    }
-
-    #[kithara::test]
-    fn eq_band_count_tracks_a_replacement_layout_before_start() {
-        let player = PlayerImpl::new(
-            PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
-                .worker(worker())
-                .session(testing::test_session())
-                .eq_layout(generate_log_spaced_bands(3))
-                .build(),
-        );
-        assert_eq!(player.eq_band_count(), 3);
-
-        player.set_eq_layout(generate_log_spaced_bands(4)).unwrap();
-        assert_eq!(player.eq_band_count(), 4);
-    }
-
-    #[kithara::test]
-    fn player_config_builder() {
-        let config = PlayerConfig::builder()
-            .sample_rate(testing::TEST_SAMPLE_RATE)
-            .worker(worker())
-            .session(testing::test_session())
-            .default_rate(0.5)
-            .crossfade_duration(2.5)
-            .prefetch_duration(7.0)
-            .max_slots(8)
-            .eq_layout(generate_log_spaced_bands(5))
-            .build();
-        assert_eq!(config.max_slots, 8);
-        assert!((config.default_rate - 0.5).abs() < f32::EPSILON);
-        assert!((config.crossfade_duration - 2.5).abs() < f32::EPSILON);
-        assert!((config.prefetch_duration - 7.0).abs() < f32::EPSILON);
-        assert_eq!(config.eq_layout.len(), 5);
-    }
-
-    #[kithara::test]
     fn player_default_rate_getter_setter() {
         let player = player();
         assert!((player.default_rate() - 1.0).abs() < f32::EPSILON);
@@ -695,44 +504,6 @@ mod tests {
         assert!((player.default_rate() - 0.75).abs() < f32::EPSILON);
         assert!((player.core.warp.stretch().speed() - 0.75).abs() < f32::EPSILON);
         assert_eq!(player.rate(), 0.0);
-    }
-
-    #[kithara::test(tokio)]
-    async fn synchronous_player_events_remain_in_order() {
-        let player = player();
-        let mut rx = player.subscribe();
-
-        player.set_volume(0.5);
-        player.set_muted(true);
-        player.set_rate(2.0);
-
-        let e1 = rx.try_recv();
-        let e2 = rx.try_recv();
-        assert!(matches!(
-            e1,
-            Ok(Envelope {
-                event: Event::Player(PlayerEvent::VolumeChanged { .. }),
-                ..
-            })
-        ));
-        assert!(matches!(
-            e2,
-            Ok(Envelope {
-                event: Event::Player(PlayerEvent::MuteChanged { .. }),
-                ..
-            })
-        ));
-        assert!(
-            rx.try_recv().is_err(),
-            "rate feedback must wait for the RT processor"
-        );
-    }
-
-    #[kithara::test(tokio)]
-    async fn player_negative_crossfade_duration_clamped() {
-        let player = player();
-        player.set_crossfade_duration(-5.0);
-        assert!((player.crossfade_duration() - 0.0).abs() < f32::EPSILON);
     }
 
     #[kithara::test]
@@ -747,9 +518,9 @@ mod tests {
     fn timestretch_is_address_stable_across_play_pause() {
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session())
+                .session(mock::session())
                 .build(),
         );
         let ptr_before = Arc::as_ptr(player.core.warp.stretch());
@@ -779,87 +550,13 @@ mod tests {
     }
 
     #[kithara::test]
-    fn position_seconds_idle_is_none() {
-        let player = player();
-        assert!(player.position_seconds().is_none());
-        assert!(player.duration_seconds().is_none());
-        assert!(!player.is_playing());
-        assert!(player.current_abr_handle().is_none());
-        assert!(player.armed_next().is_none());
-    }
-
-    #[kithara::test]
-    fn set_rate_without_rt_does_not_emit_rate_changed() {
-        let player = player();
-        let mut rx = player.subscribe();
-        player.set_rate(2.0);
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[kithara::test]
-    fn player_keeps_explicit_worker_and_shared_pools() {
-        let worker = worker();
-        let player = PlayerImpl::new(
-            PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
-                .worker(worker.clone())
-                .session(testing::test_session())
-                .build(),
-        );
-        assert!(std::ptr::eq(player.worker().pools(), worker.pools()));
-    }
-
-    #[kithara::test]
-    fn auto_advance_enabled_default_and_toggle() {
-        let player = player();
-        assert!(player.auto_advance_enabled(), "default must be on");
-        player.set_auto_advance_enabled(false);
-        assert!(!player.auto_advance_enabled());
-        player.set_auto_advance_enabled(true);
-        assert!(player.auto_advance_enabled());
-    }
-
-    #[kithara::test]
-    fn auto_advance_disabled_via_config() {
-        let player = PlayerImpl::new(
-            PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
-                .worker(worker())
-                .session(testing::test_session())
-                .auto_advance_enabled(false)
-                .build(),
-        );
-        assert!(!player.auto_advance_enabled());
-    }
-
-    #[kithara::test]
-    fn host_rejects_a_player_built_for_another_sample_rate() {
-        let mut player = PlayerImpl::new(
-            PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
-                .worker(worker())
-                .build(),
-        );
-        let binding = SessionBinding::new(Arc::new(ForeignRateSession), FOREIGN_SAMPLE_RATE);
-
-        assert!(matches!(
-            PlayerControlSource::attach_session(&mut player, binding),
-            Err(PlayError::SessionSampleRateMismatch {
-                player: 44_100,
-                session: 48_000,
-            })
-        ));
-    }
-
-    #[kithara::test]
     fn prebound_session_rejects_a_player_built_for_another_sample_rate() {
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(SessionBinding::new(
-                    Arc::new(ForeignRateSession),
-                    FOREIGN_SAMPLE_RATE,
+                .session(mock::session_at(
+                    NonZeroU32::new(48_000).expect("48000 is not zero"),
                 ))
                 .build(),
         );
@@ -871,5 +568,11 @@ mod tests {
                 session: 48_000,
             })
         ));
+    }
+
+    #[kithara::test]
+    fn send_to_slot_without_a_slot_is_an_error() {
+        let player = player();
+        assert!(player.send_to_slot(PlayerCmd::SetPaused(true)).is_err());
     }
 }
