+++ crates/kithara-play/src/player/flow/prepare.rs
@@ -147,10 +147,10 @@ mod tests {
 
     use super::*;
     use crate::{
-        PlayError, PlayWorker, PlayWorkerConfig, PlaybackResamplerBackend,
+        PlayError, PlayWorker, PlayWorkerConfig, PlaybackResamplerBackend, mock,
         player::PlayerConfig,
         resource::ResourceSrc,
-        session::{Cmd, Reply, SessionBinding, SessionDispatcher, testing},
+        session::{Cmd, Reply, SessionBinding, SessionDispatcher},
         test_pools::{TestPools, pools},
     };
 
@@ -185,16 +185,16 @@ mod tests {
     ) -> PlayerImpl<TestPools> {
         let shape = StreamShape::new(
             NonZeroU32::new(output_buffer).expect("fixture output block is non-zero"),
-            testing::TEST_SAMPLE_RATE,
+            mock::SAMPLE_RATE,
         );
         let warp = WarpConfig::builder()
             .render_quantum_frames(NonZeroUsize::new(quantum).expect("fixture quantum is non-zero"))
             .build();
         PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session_with_shape(Some(shape)))
+                .session(mock::session_with_shape(Some(shape)))
                 .warp(warp)
                 .response_budget_frames(
                     NonZeroUsize::new(response_budget).expect("fixture budget is non-zero"),
@@ -206,12 +206,12 @@ mod tests {
     #[kithara::test]
     fn prepare_config_propagates_session_consumer_wake_mode_to_audio() {
         let session = SessionBinding::new(
-            Arc::new(ImmediateSession(testing::test_session().dispatcher())),
-            testing::TEST_SAMPLE_RATE,
+            Arc::new(ImmediateSession(mock::session().dispatcher())),
+            mock::SAMPLE_RATE,
         );
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
                 .session(session)
                 .build(),
@@ -242,13 +242,13 @@ mod tests {
     fn prepare_config_sizes_default_resampling_work_to_the_output_block() {
         let shape = StreamShape::new(
             NonZeroU32::new(128).expect("test block is non-zero"),
-            testing::TEST_SAMPLE_RATE,
+            mock::SAMPLE_RATE,
         );
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session_with_shape(Some(shape)))
+                .session(mock::session_with_shape(Some(shape)))
                 .build(),
         );
 
@@ -271,7 +271,7 @@ mod tests {
     fn prepare_config_without_a_session_keeps_default_resampling_work() {
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
                 .build(),
         );
@@ -293,13 +293,13 @@ mod tests {
         config.decoder = AudioDecoderConfig::builder().resampler(explicit).build();
         let shape = StreamShape::new(
             NonZeroU32::new(128).expect("test block is non-zero"),
-            testing::TEST_SAMPLE_RATE,
+            mock::SAMPLE_RATE,
         );
         let player = PlayerImpl::new(
             PlayerConfig::builder()
-                .sample_rate(testing::TEST_SAMPLE_RATE)
+                .sample_rate(mock::SAMPLE_RATE)
                 .worker(worker())
-                .session(testing::test_session_with_shape(Some(shape)))
+                .session(mock::session_with_shape(Some(shape)))
                 .build(),
         );
