+++ crates/kithara-play/src/session/mod.rs
@@ -1,8 +1,6 @@
 //! Lower player-to-host session protocol.
 
 pub mod protocol;
-#[cfg(test)]
-pub(crate) mod testing;
 pub use protocol::{
     AllocatedSlot, Cmd, PlayerId, PlayerLevel, Reply, SessionBinding, SessionDispatcher,
     SessionError, SessionHandle, SessionSampleRate,
